Fix Coverity#1191181: incorrect sign extension
From the coverity report:
  Suspicious implicit sign extension: this->surface->pitch with type Uint16 (16
  bits, unsigned) is promoted in this->surface->pitch * this->surface->h to
  type int (32 bits, signed), then sign-extended to type unsigned long (64
  bits, unsigned). If this->surface->pitch * this->surface->h is greater than
  0x7FFFFFFF, the upper bits of the result will all be 1.

This likely never triggers in practice. But easy to fix.

+++ src/video/SDLOffScreenSurface.cc
@@ -38,7 +38,7 @@ void SDLOffScreenSurface::saveScreenshot(const std::string& filename)
 
 void SDLOffScreenSurface::clearScreen()
 {
-	memset(surface->pixels, 0, surface->pitch * surface->h);
+	memset(surface->pixels, 0, uint32_t(surface->pitch) * surface->h);
 }
 
 } // namespace openmsx
